Split TestCases to partial classes.
