+++ src/SenseNet.Security.Tests/SenseNet.Security.Tests.csproj
@@ -196,6 +196,8 @@
     <Compile Include="PermissionQueryTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TestCases.cs" />
+    <Compile Include="TestCasesAclTests.cs" />
+    <Compile Include="TestCasesMembershipTests.cs" />
     <Compile Include="TestPortal\Context.cs" />
     <Compile Include="TestPortal\EventId.cs" />
     <Compile Include="TestPortal\PermissionType.cs" />
@@ -205,6 +207,7 @@
     <Compile Include="TestPortal\TestUser.cs" />
     <Compile Include="StructureTests.cs" />
     <Compile Include="SystemStartTests.cs" />
+    <Compile Include="TestsCasesStructureTests.cs" />
     <Compile Include="Tools.cs" />
     <Compile Include="BigSystemStartTest.cs" />
   </ItemGroup>
