+++ src/SenseNet.Security.Tests/TestCasesMembershipTests.cs
@@ -1,0 +1,6 @@
+﻿namespace SenseNet.Security.Tests
+{
+    public abstract partial class TestsCases
+    {
+    }
+}
