+++ src/SenseNet.Security.Tests/TestsCasesStructureTests.cs
@@ -1,0 +1,6 @@
+﻿namespace SenseNet.Security.Tests
+{
+    public abstract partial class TestsCases
+    {
+    }
+}
